Fix some problems of AVLTree
