+++ kernel/core/AVLTree.h
@@ -1,9 +1,22 @@
-#ifndef AVLTREE
-#define AVLTREE
+/*
+ * AVLTree.h
+ *
+ *  Created on: 4 de set de 2018
+ *      Author: jpone
+ */
+
+#ifndef KERNEL_CORE_AVLTREE_H_
+#define KERNEL_CORE_AVLTREE_H_
+
 
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
+#include "portmacro.h"
+#include "EventOS.h"
+#include "list.h"
+
+#include "Log.h"
 
 /*********************************************************
 
@@ -17,17 +30,17 @@
     public types.
 
 *********************************************************/
-typedef struct tag_Info {
-	
-	portUCHAR ucValue;
-	portCHAR cMyString[10];
-	
-}ttag_info;
+typedef struct tag_Event {
 
+	xList pxSubscriberList;	/*< Event handler separated by event types. */
+	portCHAR *pcEventName;
+	portHASH_TYPE xHash;
+
+}ttag_Event;
 
 typedef struct tag_node
 {
-	ttag_info info;
+	ttag_Event Event;
 	struct tag_node *ptagRight;
 	struct tag_node *ptagLeft;
 }ttag_node;
@@ -39,12 +52,12 @@ typedef ttag_node* ttag_nodeptr;
     public operations.
 
 *********************************************************/
-ttag_nodeptr 	AVLTree_insertNode(ttag_nodeptr ptagRoot, ttag_info tagNewInfo);
-ttag_nodeptr 	AVLTree_getHandler(ttag_nodeptr ptagRoot, ttag_info tagSearchedInfo);
+ttag_nodeptr 	AVLTree_insertNode(ttag_nodeptr ptagRoot, ttag_Event tagNewInfo);
+ttag_nodeptr 	AVLTree_getHandler(ttag_nodeptr ptagRoot, ttag_Event tagSearchedInfo);
 ttag_nodeptr 	AVLTree_removeSpecificNode(ttag_nodeptr ptagRoot, ttag_nodeptr ptagNodeHandler);
-ttag_nodeptr 	AVLTree_removeNode(ttag_nodeptr ptagRoot, ttag_info tagInfo);
+ttag_nodeptr 	AVLTree_removeNode(ttag_nodeptr ptagRoot, ttag_Event tagInfo);
 ttag_nodeptr	AVLTree_clearTree(ttag_nodeptr ptagRoot);
 void			AVLTree_printTree(ttag_nodeptr ptagRoot);
 void			AVLTree_printNode(ttag_nodeptr ptagRoot);
 
-#endif
+#endif /* KERNEL_CORE_AVLTREE_H_ */
