+++ projects/EventOS_example/LPCXPRESSO/NXPLPC1769-OM13085/EA-XPR-021/Debug/EventOS_example_Debug_library.ld
@@ -5,7 +5,7 @@
  * Generated linker script file for LPC1769
  * Created from library.ldt by FMCreateLinkLibraries
  * Using Freemarker v2.3.23
- * LPCXpresso v8.2.2 [Build 650] [2016-09-09]  on 18/09/2018 20:16:14
+ * LPCXpresso v8.2.2 [Build 650] [2016-09-09]  on 18/09/2018 21:01:31
  */
 
 GROUP (
