+++ projects/EventOS_example/LPCXPRESSO/NXPLPC1769-OM13085/EA-XPR-021/Debug/Kernel/Core/AVLTree.d
@@ -1,17 +1,15 @@
-Kernel/Core/AVLTree.o Kernel/Core/AVLTree.d: ../Kernel/Core/AVLTree.c \
- ../Kernel/Core/AVLTree.h \
+Kernel/Core/AVLTree.o Kernel/Core/AVLTree.d: \
+ D:/JoaoDoAviao/Inatel/TCC/repositorio/EventOS/kernel/core/AVLTree.c \
+ D:/JoaoDoAviao/Inatel/TCC/repositorio/EventOS/kernel/core/AVLTree.h \
  ../../../../../../kernel/portable/device/NXP/LPC17xx/portmacro.h \
  ../../../../../../kernel/config/EventOS.h \
  ../../../../../../kernel/config/EventOSConfig.h \
  ../../../../../../kernel/config/portable.h \
  ../../../../../../kernel/config/..\portable\device\NXP\LPC17xx\portmacro.h \
- ../../../../../../kernel/core/event.h \
- ../../../../../../kernel/config/portable.h \
- ../../../../../../kernel/core/list.h \
- ../../../../../../kernel/core/list.h \
+ D:/JoaoDoAviao/Inatel/TCC/repositorio/EventOS/kernel/core/list.h \
  ../../../../../../middlewares/Log/Log.h
 
-../Kernel/Core/AVLTree.h:
+D:/JoaoDoAviao/Inatel/TCC/repositorio/EventOS/kernel/core/AVLTree.h:
 
 ../../../../../../kernel/portable/device/NXP/LPC17xx/portmacro.h:
 
@@ -23,12 +21,6 @@ Kernel/Core/AVLTree.o Kernel/Core/AVLTree.d: ../Kernel/Core/AVLTree.c \
 
 ../../../../../../kernel/config/..\portable\device\NXP\LPC17xx\portmacro.h:
 
-../../../../../../kernel/core/event.h:
-
-../../../../../../kernel/config/portable.h:
-
-../../../../../../kernel/core/list.h:
-
-../../../../../../kernel/core/list.h:
+D:/JoaoDoAviao/Inatel/TCC/repositorio/EventOS/kernel/core/list.h:
 
 ../../../../../../middlewares/Log/Log.h:
